requirements.txt
torch
torchvision
timm
numpy
scikit-learn
pandas
matplotlib
tqdm
Pillow
<<<<<<< HEAD
git+https://github.com/openai/CLIP.git
=======
requests


# Transformers (DINOv2)
transformers

# CLIP (OpenAI GitHub version)
git+https://github.com/openai/CLIP.git

# Optional: Uncomment for specific versions / CUDA compatibility
# torch==2.2.2+cu118
# torchvision==0.17.2+cu118
# transformers==4.37.2
# scikit-learn==1.4.2
# timm==0.9.12
# Pillow==11.2.0
# matplotlib==3.8.4
# tqdm==4.66.4
>>>>>>> 30k_dataset
